Cleanup MSVS build warnings by adding appropriate casts and suppressions
Cleanup MSVS build warnings by adding appropriate casts and suppressions.

The warnings that are now suppressed:

    # /wd4251 Qt warning about dll interface
    # /wd4275 Qt warning about dll interface
    # /wd4018 signed/unsigned mismatch
    # /wd4244 conversion from double to float
    # /wd4305 truncation from double to float
    # /wd4267 conversion from size_t to int (Qt uses ints extensively)

+++ graphics/bitmap.h
@@ -191,9 +191,9 @@ namespace gfx
     {
         const float a = rgba.a / 255.0f;
         // bake the alpha in the color channels
-        r = rgba.r * a;
-        g = rgba.g * a;
-        b = rgba.b * a;
+        r = u8(rgba.r * a);
+        g = u8(rgba.g * a);
+        b = u8(rgba.b * a);
     }
 
     // Grayscale methods that depend on RGB/A come here.
@@ -205,8 +205,8 @@ namespace gfx
         const float y =
             0.2126f * rgb.r +
             0.7252f * rgb.g +
-            0.0722  * rgb.b;
-        r = y / 255.0f;
+            0.0722f * rgb.b;
+        r = u8(y / 255.0f);
     }
     inline Grayscale::Grayscale(const RGBA& rgba)
     {
@@ -217,8 +217,8 @@ namespace gfx
         const float y =
             0.2126f * rgba.r +
             0.7252f * rgba.g +
-            0.0722  * rgba.b;
-        r = (y * a) / 255.0f;
+            0.0722f * rgba.b;
+        r = u8((y * a) / 255.0f);
     }
 
     static_assert(sizeof(Grayscale) == 1,
